Separated shared code into base class.
